Former-commit-id: 2d7f3193f8fef35d9046485cbf83ab947a65d826

+++ htdocs/bimpcommercial/objects/BimpComm.class.php
@@ -330,7 +330,7 @@ class BimpComm extends BimpDolObject
 
     public function useEntrepot()
     {
-        return BimpCore::getConf("USE_ENTREPOT");
+        return (int) BimpCore::getConf("USE_ENTREPOT");
     }
 
     // Getters array:
